pyproject.toml
[build-system]
requires = ["hatchling", "requests", "appdirs", "hatch-build-scripts"]
build-backend = "hatchling.build"

[project]
name = "jpyjs9"
description = 'Running js9 on inside jupyter'
readme = "README.md"
requires-python = ">=3.7"
license = "MIT"
keywords = []
authors = [
  { name = "Abdu Zoghbi", email = "a.zoghbi@nasa.gov" },
]

dependencies = [
    "jupyter-server-proxy @ git+https://github.com/jupyterhub/jupyter-server-proxy",
    "python-socketio[client]",
    "socketIO-client",
    "sidecar",
    "pyjs9 @ git+https://github.com/duytnguyendtn/pyjs9.git@jupyterjs9",
    "appdirs"
]
dynamic = ["version"]

[project.entry-points.jupyter_serverproxy_servers]
js9 = "jpyjs9:js9_web_server"
js9Helper = "jpyjs9:js9_helper_server"

[project.urls]
Source = "https://github.com/zoghbi-a/jpyjs9"

[tool.hatch.version]
path = "jpyjs9/__init__.py"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel.shared-data]
"jpyjs9.json" = "etc/jupyter/jupyter_server_config.d/jpyjs9.json"

[tool.hatch.build.targets.sdist]
exclude = [".git", ".ipynb_checkpoints"]

[[tool.hatch.build.hooks.build-scripts.scripts]]
commands = [
    "python clone_js9.py"
]
artifacts = []